Cli and server split
