+++ cmd/commands/leave.go
@@ -63,7 +63,7 @@ func (c *LeaveCommand) Run(args []string) int {
 
 	defer connection.Close()
 
-	client := proto.NewRequesterClient(connection)
+	client := proto.NewCLIClient(connection)
 
 	r, err := client.Leave(context.Background(), &proto.PulseLeave{})
