+++ src/client.go
@@ -22,28 +22,27 @@ import (
 	"google.golang.org/grpc/credentials"
 	"github.com/coreos/go-log/log"
 	p "github.com/Syleron/PulseHA/proto"
-	"context"
 	"errors"
 )
 
 type Client struct {
 	//State      PulseState
 	Connection *grpc.ClientConn
-	Requester  p.RequesterClient
+	Requester  p.ServerClient
 }
 
 /**
- *
+
  */
 func (c *Client) GetFuncBroadcastList() (map[string]interface{}) {
 	funcList := map[string]interface{} {
-		"SendJoin": c.SendJoin,
+		//"SendJoin": c.SendJoin,
 	}
 	return funcList
 }
 
 /**
- *
+
  */
 func (c *Client) Setup() {
 }
@@ -69,107 +68,107 @@ func (c *Client) Connect(ip, port, hostname string) (error) {
 		log.Errorf("GRPC client connection error: ", err)
 		return err
 	}
-	c.Requester = p.NewRequesterClient(c.Connection)
+	c.Requester = p.NewServerClient(c.Connection)
 	return nil
 }
 
 /**
- *
+	Close the client connection
  */
 func (c *Client) Close() {
 	log.Debug("Client:Close() Connection closed")
 	c.Connection.Close()
 }
 
-// Senders. Consider moving these into their own file
-
-/**
- *
- */
-func (c *Client) SendCheck(data *p.HealthCheckRequest) (*p.HealthCheckResponse, error) {
-	log.Debug("Client:SendCheck() Sending GRPC Check")
-	r, err := c.Requester.Check(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendJoin(data *p.PulseJoin) (*p.PulseJoin, error) {
-	log.Debug("Client:SendJoin() Sending GRPC Join")
-	r, err := c.Requester.Join(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendLeave(data *p.PulseLeave) (*p.PulseLeave, error) {
-	log.Debug("Client:SendLeave() Sending GRPC Leave")
-	r, err := c.Requester.Leave(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendGroupNew(data *p.PulseGroupNew) (*p.PulseGroupNew, error) {
-	log.Debug("Client:SendGroupNew() Sending GRPC NewGroup")
-	r, err := c.Requester.NewGroup(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendGroupDelete(data *p.PulseGroupDelete) (*p.PulseGroupDelete, error) {
-	log.Debug("Client:SendGroupDelete() Sending GRPC DeleteGroup")
-	r, err := c.Requester.DeleteGroup(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendGroupIPAdd(data *p.PulseGroupAdd) (*p.PulseGroupAdd, error) {
-	log.Debug("Client:SendGroupIPAdd() Sending GRPC GroupIPAdd")
-	r, err := c.Requester.GroupIPAdd(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendGroupIPRemove(data *p.PulseGroupRemove) (*p.PulseGroupRemove, error) {
-	log.Debug("Client:SendGroupIPRemove() Sending GRPC GroupIPRemove")
-	r, err := c.Requester.GroupIPRemove(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendGroupAssign(data *p.PulseGroupAssign) (*p.PulseGroupAssign, error) {
-	log.Debug("Client:SendGroupAssign() Sending GRPC GroupAssign")
-	r, err := c.Requester.GroupAssign(context.Background(), data)
-
-	return r, err
-}
-
-/**
- *
- */
-func (c *Client) SendGroupUnassign(data *p.PulseGroupUnassign) (*p.PulseGroupUnassign, error) {
-	log.Debug("Client:SendGroupUnassign() Sending GRPC GroupUnassign")
-	r, err := c.Requester.GroupUnassign(context.Background(), data)
-
-	return r, err
-}
+//// Senders. Consider moving these into their own file
+//
+///**
+// *
+// */
+//func (c *Client) SendCheck(data *p.HealthCheckRequest) (*p.HealthCheckResponse, error) {
+//	log.Debug("Client:SendCheck() Sending GRPC Check")
+//	r, err := c.Requester.Check(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendJoin(data *p.PulseJoin) (*p.PulseJoin, error) {
+//	log.Debug("Client:SendJoin() Sending GRPC Join")
+//	r, err := c.Requester.Join(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendLeave(data *p.PulseLeave) (*p.PulseLeave, error) {
+//	log.Debug("Client:SendLeave() Sending GRPC Leave")
+//	r, err := c.Requester.Leave(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendGroupNew(data *p.PulseGroupNew) (*p.PulseGroupNew, error) {
+//	log.Debug("Client:SendGroupNew() Sending GRPC NewGroup")
+//	r, err := c.Requester.NewGroup(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendGroupDelete(data *p.PulseGroupDelete) (*p.PulseGroupDelete, error) {
+//	log.Debug("Client:SendGroupDelete() Sending GRPC DeleteGroup")
+//	r, err := c.Requester.DeleteGroup(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendGroupIPAdd(data *p.PulseGroupAdd) (*p.PulseGroupAdd, error) {
+//	log.Debug("Client:SendGroupIPAdd() Sending GRPC GroupIPAdd")
+//	r, err := c.Requester.GroupIPAdd(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendGroupIPRemove(data *p.PulseGroupRemove) (*p.PulseGroupRemove, error) {
+//	log.Debug("Client:SendGroupIPRemove() Sending GRPC GroupIPRemove")
+//	r, err := c.Requester.GroupIPRemove(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendGroupAssign(data *p.PulseGroupAssign) (*p.PulseGroupAssign, error) {
+//	log.Debug("Client:SendGroupAssign() Sending GRPC GroupAssign")
+//	r, err := c.Requester.GroupAssign(context.Background(), data)
+//
+//	return r, err
+//}
+//
+///**
+// *
+// */
+//func (c *Client) SendGroupUnassign(data *p.PulseGroupUnassign) (*p.PulseGroupUnassign, error) {
+//	log.Debug("Client:SendGroupUnassign() Sending GRPC GroupUnassign")
+//	r, err := c.Requester.GroupUnassign(context.Background(), data)
+//
+//	return r, err
+//}
