+++ src/main.go
@@ -45,6 +45,7 @@ var (
  */
 type Pulse struct {
 	Server *Server
+	CLI *CLIServer
 }
 
 /**
@@ -58,8 +59,10 @@ func createPulse() *Pulse {
 	// Create the Pulse object
 	pulse := &Pulse{
 		Server: &Server{
-			Memberlist: &Memberlist{
-			},
+			Memberlist: &Memberlist{},
+		},
+		CLI: &CLIServer{
+			Memberlist: &Memberlist{},
 		},
 	}
 	return pulse
@@ -88,7 +91,7 @@ func main() {
 	var wg sync.WaitGroup
 	wg.Add(1)
 	// Setup cli
-	go pulse.Server.SetupCLI()
+	go pulse.CLI.Setup()
 	// Setup server
 	go pulse.Server.Setup()
 	wg.Wait()
